Facade id change from int to long

+++ src/main/java/facade/CompanyFacade.java
@@ -60,7 +60,7 @@ public class CompanyFacade implements ICompanyFacade
     }
 
     @Override
-    public Company deleteCompany(int id)
+    public Company deleteCompany(long id)
     {
         try
         {
